make generator faster by processing files in parallel
